add logic to turn leader back into follower and allocate seperate vector for followers

+++ src/election/mod.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
 
-pub fn elect(n: u32) -> u32 {
+pub fn elect(n: usize) -> usize {
     println!("Electing...");
     rand::thread_rng().gen_range(0..n)
 }
